+++ changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## Added
-- Added the `newAggregate()` function to customize the aggregate query builder ([#11](https://github.com/Jeroen-G/Explorer/pull/11))
+- Added the `newAggregate()` function to customize the aggregate query builder
 
 ### Fixed
 - Boosts may now completely be left out